// ready_made_angular/src/app/models/job.model.ts
export interface Job {
    job_id: string;
    product_code?: string;
    product_size?: string;
    is_selected: boolean;

    order_id?: string;
    order_serial?: number;
    order_no?: number;
    rm_id?: number;
    pieces?: number;
    job_date?: number;
    delivery_date?: number;
    expected_gold?: number;
    p_loss?: number;
    tr_date?: number;
    tr_time?: string;
    status?: number;
    emp_id?: number;
    price_method?: string;
    price_code?: string;
    price?: number;
    gold_send?: number;
    dal_send?: number;
    pan_send?: number;
    bronze_send?: number;
    copper_send?: number;
    gold_returned?: number;
    dal_returned?: number;
    pan_returned?: number;
    bronze_returned?: number;
    nitrick_returned?: number;
    copper_return?: number;
    product_wt?: number;
    comments?: string;
    dal_id?: number;
    pan_id?: number;
    bronze_id?: number;
    copper_id?: number | null;
    markup_value?: number;
    in_stock?: number;
    ip?: string;
}